fix `finalWrite` for split fixed-size cursors
When a `delayFixedSizeWrite` cursor is split across pages, it can't be
written using a simple `copyMem`. Fix `finalWrite` to use the correct
`writeBytesToCursor` implementation instead, which still uses `copyMem`
for the cases that worked fine, so should also have same performance.
`finalize` functions had to be moved down to use `writeBytesToCursor`.