PageBuffers revamp (#71)
Reimplement `PageBuffers` primtives simplify its usage in streams. In
particular, buffers now follow the established
`prepare`/`commit`/`consume` pattern for streaming also seen in chronos
and asio.

In short, each time a buffer is needed, `prepare` is called to allocate
an appropriate buffer - once the data has been written, `commit`
notifies the buffer of how many bytes were actually written.
Subsequently, `consume` gives a reader access to the committed bytes.

In addition to `prepare`, `reserve` can be used for delayed-write
patterns, moving the tracking of both variable-sized and fixed-size
delayed writes directly into PageBuffers.

This is the first round of cleanups that aims to address bugs and issues
with how delayed writes are tracked, as well as reduce the complexity of
the implementation while maintaining its performance profile.

With this change, `PageBuffers` becomes usable as a general-purpose
paged buffer outside of its usage in faststreams.

* refactor primitive buffer operations into prepare/commit/consume
* document PageBuffers
* deprecate most of the old PageBuffers API - will be removed at some
point in the near future
* move reservation logic from OutputStream to PageBuffers, fixing
several bugs in the process - both fixed and variable-sized reservations
now use the same logic simplifying the implementation
* introduce post-factum buffer overrun checks for PageSpan when
switching spans
* delay the allocation of new pages after flush such that if no write
happens, no page is allocated